🔨  Switch from gastby-source-git to gatsby-source-file
