+++ gatsby-node.js
@@ -29,7 +29,6 @@ const elasticlunr = require(`elasticlunr`)
 const { GraphQLJSONObject } = require("graphql-type-json")
 const converter = require("widdershins")
 
-const environment = process.env.NODE_ENV || "development"
 const openApiSearchDocs = []
 
 const pages = []
@@ -86,30 +85,14 @@ exports.onCreateNode = ({ node, getNode, actions }) => {
   const { createNodeField } = actions
   if (node.internal.type === `Mdx`) {
     let slug = ""
-    switch (environment) {
-      case "production":
-        if (node.fileAbsolutePath.lastIndexOf("gatsby-source-git/") > -1) {
-          slug = node.fileAbsolutePath.substring(
-            node.fileAbsolutePath.lastIndexOf("gatsby-source-git/") + 18
-          )
-        } else if (node.frontmatter.path) {
-          slug = node.frontmatter.path
-        }
-        break
-      case "development":
-        const localFilePath = path.relative(__dirname, node.fileAbsolutePath)
-        const directories = localFilePath.split(path.sep)
-
-        // Remove src/content prefix from slug
-        if (localFilePath.startsWith("src/content")) {
-          // Remove src
-          directories.shift()
-          // Remove content
-          directories.shift()
-        }
-
-        slug = path.join("/", ...directories)
-        break
+    if (node.frontmatter.path) {
+      slug = node.frontmatter.path
+    } else {
+      const localFilePath = path.relative(
+        `${process.env.LOCAL_PROJECT_DIRECTORY}`,
+        node.fileAbsolutePath
+      )
+      slug = path.join("/", ...localFilePath.split(path.sep))
     }
 
     if (
@@ -177,20 +160,6 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
             }
           }
         }
-        allGithubContributors {
-          edges {
-            node {
-              id
-              contributors {
-                date
-                login
-                name
-                avatarUrl
-              }
-              path
-            }
-          }
-        }
         parliamentNavigation {
           pages
         }
@@ -207,7 +176,8 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
   result.data.allMdx.edges.map(post => {
     post.node.fields.slug.includes("blog/") ? posts.push(post) : docs.push(post)
   })
-  const contributors = result.data.allGithubContributors.edges
+  // const contributors = result.data.allGithubContributors.edges
+  const contributors = []
   const parliamentNavigation = result.data.parliamentNavigation
 
   if (posts.length > 0) {
@@ -266,7 +236,7 @@ exports.createPages = async ({ actions, graphql, reporter }) => {
       postsNav.pages.push({
         importedFileName: "posts",
         pages: [],
-        path: `${gitPathPrefix}/${post.node.fields.slug}`,
+        path: `${gitPathPrefix}/${post.node.fields.slug}`.replace(/\/\//g, "/"),
         title: post.node.frontmatter.title,
       })
 
@@ -507,29 +477,12 @@ const createOpenApiPage = async (
   if (object && (object.swagger || object.openapi)) {
     let slug = filepath
 
-    switch (environment) {
-      case "production":
-        if (filepath.lastIndexOf("gatsby-source-git/") > -1) {
-          slug = filepath.substring(
-            filepath.lastIndexOf("gatsby-source-git/") + 18
-          )
-        }
-        break
-      case "development":
-        const localFilePath = path.relative(__dirname, filepath)
-        const directories = localFilePath.split(path.sep)
-
-        // Remove src/content prefix from slug
-        if (localFilePath.startsWith("src/content")) {
-          // Remove src
-          directories.shift()
-          // Remove content
-          directories.shift()
-        }
-
-        slug = path.join("/", ...directories)
-        break
-    }
+    const localFilePath = path.relative(
+      `${process.env.LOCAL_PROJECT_DIRECTORY}`,
+      filepath
+    )
+    const directories = localFilePath.split(path.sep)
+    slug = path.join("/", ...directories)
 
     try {
       const targets = [
