--- plugins/link-local-project/package.json
@@ -1,1 +1,0 @@
-{}
